Move extractor implementations to package
