+++ extractor/extractor_test.go
@@ -1,4 +1,4 @@
-package empire
+package extractor
 
 import (
 	"archive/tar"
